Added jit-compatible MTOV limitation to constructed family
Before, the M, R and Lambda curves did not terminate if MTOV was reached. How this is handled in e.g. NMMA is not jit-compatible. In this implementation, which is however quite cumbersome, there is a jit-friendly way to limit the curves up to MTOV. However, this involves duplicating the first entry of the curves to preserve the shape of the arrays. When creating interpolation representations, therefore we first need to call jnp.unique on the curves. NOTE: this assumes that the M curve increases up to a point and potentially decreases after that point. In case the EOS has some weird features and the M curve increases again later on, this function will return weird results.